+++ src/main/java/org/lightjason/agentspeak/action/builtin/graph/CDegreeSingle.java
@@ -58,7 +58,7 @@ public final class CDegreeSingle extends IApplySingle
     {
         p_return.add(
             CRawTerm.from(
-                (long) p_graph.degree( p_window.get( 0 ).raw() )
+                (double) p_graph.degree( p_window.get( 0 ).raw() )
             )
         );
     }
