+++ src/main/java/org/lightjason/agentspeak/action/builtin/graph/CIncidentCountMultiple.java
@@ -57,7 +57,7 @@ public final class CIncidentCountMultiple extends IApplyMultiple
     {
         p_return.add(
             CRawTerm.from(
-                (long) p_graph.getIncidentCount( p_window.get( 0 ).raw() )
+                (double) p_graph.getIncidentCount( p_window.get( 0 ).raw() )
             )
         );
     }
