+++ src/main/java/org/lightjason/agentspeak/action/builtin/graph/CSuccessorCountMultiple.java
@@ -53,7 +53,7 @@ public final class CSuccessorCountMultiple extends IApplyMultiple
     {
         p_return.add(
             CRawTerm.from(
-                (long) p_graph.getSuccessorCount( p_window.get( 0 ).raw() )
+                (double) p_graph.getSuccessorCount( p_window.get( 0 ).raw() )
             )
         );
     }
